+++ src/main/java/com/growith/tailo/notification/entity/Notification.java
@@ -1,0 +1,43 @@
+package com.growith.tailo.notification.entity;
+
+import com.growith.tailo.member.entity.Member;
+import com.growith.tailo.notification.enums.NotificationType;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDateTime;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Table(name = "notifications")
+public class Notification {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne
+    @JoinColumn(name = "receiver_id")
+    private Member receiver;
+
+    @ManyToOne
+    @JoinColumn(name = "sender_id")
+    private Member sender;
+
+    @Enumerated(EnumType.STRING)
+    private NotificationType type;
+
+    private LocalDateTime createdAt;
+
+    // getters and setters
+}
